Move Env to components
