+++ bin/debug.php
@@ -3,7 +3,7 @@
 require_once __DIR__ . '/../vendor/autoload.php';
 
 use pjpawel\LightApi\Component\Debug\KernelDebugger;
-use pjpawel\LightApi\Env;
+use pjpawel\LightApi\Component\Env;
 use pjpawel\LightApi\Kernel;
