+++ test/Component/EnvTest.php
@@ -1,26 +1,27 @@
 <?php
 
-namespace pjpawel\LightApi\Test;
+namespace pjpawel\LightApi\Test\Component;
 
 use PHPUnit\Framework\TestCase;
-use pjpawel\LightApi\Env;
+use pjpawel\LightApi\Component\Env;
+use pjpawel\LightApi\Test\resources;
 
 /**
- * @covers \pjpawel\LightApi\Env
+ * @covers \pjpawel\LightApi\Component\Env
  */
 class EnvTest extends TestCase
 {
 
     /**
-     * @covers \pjpawel\LightApi\Env::getConfigFromEnv
+     * @covers \pjpawel\LightApi\Component\Env::getConfigFromEnv
      */
     public function testGetConfigFromEnv()
     {
-        $dir = __DIR__ . '/resources/config/base_config/';
+        $dir = __DIR__ . '/../resources/config/base_config/';
         $config = Env::getConfigFromEnv($dir);
         $this->assertEquals(
             [
-                'projectDir' => realpath(__DIR__ . '/../'),
+                'projectDir' => realpath(__DIR__ . '/../../'),
                 'env'=>'test',
                 'debug' => true,
                 'trustedIPs' => [],
@@ -29,11 +30,11 @@ class EnvTest extends TestCase
                 'container' => [
                     resources\classes\Logger::class => []
                 ],
-                'services' => realpath(__DIR__ . '/resources/classes/')
+                'services' => realpath(__DIR__ . '/../resources/classes/')
             ],
             $config);
 
-        $dir = __DIR__ . '/resources/config/config_with_local/';
+        $dir = __DIR__ . '/../resources/config/config_with_local/';
         $config = Env::getConfigFromEnv($dir);
         $this->assertEquals(
             [
@@ -45,17 +46,17 @@ class EnvTest extends TestCase
                 'container' => [
                     resources\classes\Logger::class => []
                 ],
-                'services' => realpath(__DIR__ . '/resources/classes/')
+                'services' => realpath(__DIR__ . '/../resources/classes/')
             ],
             $config);
     }
 
     /**
-     * @covers \pjpawel\LightApi\Env::loadConfigFile
+     * @covers \pjpawel\LightApi\Component\Env::loadConfigFile
      */
     public function testLoadConfigFile()
     {
-        $filename = __DIR__ . '/resources/config/base_config/env.php';
+        $filename = __DIR__ . '/../resources/config/base_config/env.php';
         $config = Env::loadConfigFile($filename);
         $this->assertIsArray($config);
         $this->assertEquals(['env'=>'test', 'debug' => true], $config);
