+++ test/KernelTest.php
@@ -3,8 +3,8 @@
 namespace pjpawel\LightApi\Test;
 
 use PHPUnit\Framework\TestCase;
+use pjpawel\LightApi\Component\Env;
 use pjpawel\LightApi\Container\ContainerLoader;
-use pjpawel\LightApi\Env;
 use pjpawel\LightApi\Kernel;
 use pjpawel\LightApi\Test\resources\classes\Logger;
 use ReflectionClass;
